Revert "phy: ti: phy-gmii-sel: fix inverted rgmii_id bit"
In cab06d8, rgmii_id was incorrectly assessed as inverted. In device
tree, phy-mode bindings "rgmii-id" and "rgmii-txid" refer to delays
added internally by the PHY, not by the MAC or the layout. The bit in
question was previously used on some CPSW implementations to turn on (0)
or off (1) the *MAC* delay. Therefore, the previous implementation was
proper and the change should be reverted.

This reverts commit cab06d80539aa8502e93f5716d9dda42f25acad9.

Signed-off-by: Ken Sloat <ken.s@variscite.com>