am62x_var_som: Set wm8904 to properly use audio_refclk1
The AUDIO_EXT_REFCLK1 pin from the MCASP peripheral is used to source
the MCLK to the external audio codec. With recent kernel changes, this
clock output has now been defined in the kernel and so we should use the
associated clock node as the provider. Without this change, the
microphone audio is distorted.

Signed-off-by: Ken Sloat <ken.s@variscite.com>