am62_var_som: Change display related modules from LKM to builtin
In the Kirkstone release, TI changed the display related modules to be m
rather than y in the config. This results in the Linux penguin not being
displayed by both policy and also due to resource availability at the
time the penguin should be loaded. To restore the penguin, change the
modules back to builtins.

Signed-off-by: Ken Sloat <ken.s@variscite.com>